rename sshsecret to sshcredential and update files to reflect change.
